Repositories and Services Crud functionality massive commit apologies

+++ src/main/java/dev/n1t/appointments/dto/AppointmentRegistrationDTO.java
@@ -18,7 +18,7 @@ public class AppointmentRegistrationDTO {
 
     private Long appointmentDateTime;
 
-    private Long bankerId;
+    private int bankerId;
 
     private Long serviceTypeId;
